backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis & Caching
redis==5.0.1
aioredis==2.0.1

# Celery for background tasks
celery==5.3.4
flower==2.0.1

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Authentication
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# MikroTik RouterOS API
routeros-api==0.17.0

# gRPC for Starlink
grpcio==1.59.3
grpcio-tools==1.59.3
protobuf==4.25.1

# Data validation
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7
structlog==23.2.0

# Security
python-jose[cryptography]==3.3.0
cryptography==41.0.7

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
pytz==2023.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
